Move all the MessageBuilders into a separate directory
